Develop

+++ sdk/src/main/java/com/mirror/sdk/request/ReqTransSPLToken.java
@@ -4,5 +4,5 @@ public class ReqTransSPLToken {
     public String toPublickey;
     public float amount;
     public String token_mint;
-    public float decimals;
+    public int decimals;
 }
